Implement token limiters and a selector
Specifically, the PlaceOfOriginSelector, LastRequestLimiter, and
TimeSpanLimiter are all implemented in this revision.

These required definitions for an Api Request, a request history
provider, and a redefinition of ITokenLimiter to accept a history
rather than an access token.

In addition, everything is now a generic. The type for the generic
represents the type of the user identifier on the access token.
I've made it generic in lieu of specifying a concrete type to make
the library more flexible. In practice this makes the code a bit
uglier everywhere and wouldn't be necessary if it's only used for
a single API where the id type is known.